+++ src/app/api/auth/[...nextauth]/authOptions.ts
@@ -1,0 +1,60 @@
+import { NextAuthOptions } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+
+const authOptions: NextAuthOptions = {
+  pages: {
+    signIn: '/signin',
+  },
+  providers: [
+    CredentialsProvider({
+      name: 'common',
+      credentials: {
+        email: {
+          label: 'email',
+          type: 'text',
+        },
+        password: { label: 'Password', type: 'password' },
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials.password) return null;
+
+        try {
+          const res = await fetch('http://localhost:8080/user/signin', {
+            method: 'post',
+            body: JSON.stringify({
+              email: credentials.email,
+              password: credentials.password,
+            }),
+            headers: {
+              'Content-Type': 'application/json',
+              Accept: 'application/json',
+            },
+          });
+          const { data } = await res.json();
+          return {
+            id: data.token,
+            token: data.token,
+          };
+        } catch (error) {
+          return null;
+        }
+      },
+    }),
+  ],
+  callbacks: {
+    async session({ session, token }) {
+      // eslint-disable-next-line no-param-reassign
+      session.token = token.token;
+      return session;
+    },
+    async jwt({ token, user }) {
+      if (user) {
+        // eslint-disable-next-line no-param-reassign
+        token.token = user.token;
+      }
+      return token;
+    },
+  },
+};
+
+export default authOptions;
